+++ NGLess/StandardModules/NGLStdlib.hs
@@ -1,8 +1,7 @@
-{- Copyright 2015-2016 NGLess Authors
+{- Copyright 2015-2022 NGLess Authors
  - License: MIT
  -}
 
-{-# LANGUAGE TupleSections, OverloadedStrings #-}
 module StandardModules.NGLStdlib
     ( loadStdlibModules
     ) where
@@ -15,6 +14,7 @@ import qualified StandardModules.Mocat as Mocat
 import qualified StandardModules.Soap as Soap
 import qualified StandardModules.Minimap2 as Minimap2
 import qualified StandardModules.Parallel as Parallel
+import qualified StandardModules.Motus as Motus1
 import qualified ExternalModules as Ext
 
 import Modules
@@ -31,5 +31,6 @@ loadStdlibModules = mapM loadModules1
         loadModules1 (ModInfo "parallel" version) = Parallel.loadModule version
         loadModules1 (ModInfo "soap" version) = Soap.loadModule version
         loadModules1 (ModInfo "minimap2" version) = Minimap2.loadModule version
+        loadModules1 (ModInfo "motus" version) = Motus1.loadModule version
         loadModules1 minfo = Ext.loadModule minfo